# https://github.com/r-lib/pkgdown/issues/2704

Calculate 2012 CAD2 Basic PTP for obstructive CAD

This function returns a patient's pre-test probability (PTP) of obstructive coronary artery disease based on the 2012 CAD Consortium 2 (CAD2) basic model.

Usage

calculate_cad2_2012_basic_ptp(
  age,
  sex,
  chest_pain_type,
  label_sex_male = c("male"),
  label_sex_female = c("female"),
  label_sex_unknown = c(NA, NaN),
  label_cpt_nonanginal = c("nonanginal"),
  label_cpt_atypical = c("atypical"),
  label_cpt_typical = c("typical"),
  label_cpt_unknown = c(NA, NaN)
)

Arguments

age

Input numeric value to indicate the age of the patient in years.

sex

The value of variable in the parameters label_sex_male, label_sex_female and label_sex_unknown.

chest_pain_type

The value of variable in the parameters, label_cpt_nonanginal, label_cpt_atypical, label_cpt_typical and label_cpt_unknown.

label_sex_male

Label(s) for definition(s) of male sex. Default: c("male")

label_sex_female

Label(s) for definition(s) of female sex. Default: c("female")

label_sex_unknown

Label(s) for definition(s) of missing sex. Default: c(NA, NaN)

label_cpt_nonanginal

Label(s) for patient having nonanginal or non-specific chest pain. Default: c("nonanginal")

label_cpt_atypical

Label(s) for patient having atypical chest pain. Default: c("atypical")

label_cpt_typical

Label(s) for patient having typical chest pain. Default: c("typical")

label_cpt_unknown

Label(s) for patient having unknown chest pain type symptoms. Default: c(NA, NaN)

Value

A numeric value representing the patient's PTP for obstructive CAD based on the 2012 CAD Consortium 2 (CAD2) basic model.

Details

The predictive model is based on patients from 18 hospitals in Europe and the United States.

Examples

# 40 year old female with typical chest pain
calculate_cad2_2012_basic_ptp(
    age = 40,
    sex = "female",
    chest_pain_type = "typical"
)
#> [1] 0.0815104