# https://github.com/r-lib/pkgdown/issues/2704

Calculate 2015 CONFIRM Risk Score for obstructive CAD

This function returns a patient's risk score for obstructive coronary artery disease based on the 2015 CONFIRM Risk Score.

Usage

calculate_confirm_2015_ptp(
  age,
  sex,
  have_typical_chest_pain,
  have_diabetes,
  have_hypertension,
  have_family_history,
  is_current_smoker,
  max_na_num_of_rf = 0,
  output = c("text", "percentage")
)

Arguments

age

Input numeric value to indicate the age of the patient.

sex

Input characters (female, male) to indicate the sex of the patient.

female

male

have_typical_chest_pain

Input characters (no, yes) to indicate if the patient has typical chest pain.

no stands for not having typical chest pain.

yes stands for having typical chest pain.

have_diabetes

Input characters (no, yes) to indicate if the patient only has diabetes.

no stands for not having diabetes.

yes stands for having diabetes.

have_hypertension

Input characters (no, yes) to indicate if the patient only has hypertension.

no stands for not having hypertension.

yes stands for having a hypertension.

have_family_history

Input characters (no, yes) to indicate if the patient only has a family history of CAD.

no stands for not having a family history of CAD.

yes stands for having a family history of CAD.

is_current_smoker

Input characters (no, yes) to indicate if the patient is a current smoker.

no stands for patient is a current smoker.

yes stands for patient is a not current smoker (past or non-smoker).

max_na_num_of_rf

Input integer 0 to 5 to indicate the maximum number of missing risk factors to tolerate before outputting an NA. Default: 0

output

Input text to indicate the how pre-test probability results be expressed Default: c("text", "percentage")

text means the PTP will be expressed as a probability in text (0 to > 82.4).

percentage means the PTP will be expressed as percentage text (0-100%).

Value

A numeric value representing the patient's risk score for obstructive CAD based on the 2015 CONFIRM Risk Score.

Details

The predictive model is based on CCTA images from 9093 patients from Phase I of the Coronary CT Angiography EvaluatioN For Clinical Outcomes: An InteRnational Multicenter (CONFIRM) registry.

Examples

# 30 years old male current smoker with typical chest pain
calculate_confirm_2015_ptp(
  age = 30,
  sex = "male",
  have_typical_chest_pain = "yes",
  have_diabetes = "no",
  have_hypertension = "no",
  have_family_history = "no",
  is_current_smoker = "yes",
  max_na_num_of_rf = 0,
  output = "percentage"
)
#> [1] "5.5%"