# https://github.com/r-lib/pkgdown/issues/2704

Calculate ESC 2019 PTP for obstructive CAD

This function returns a patient's pre-test Probability (PTP) of obstructive coronary artery disease (CAD) based on the European Society of Cardiology (ESC) 2019 guidelines.

Usage

calculate_esc_2019_ptp(
  age,
  sex,
  have_dyspnoea,
  chest_pain_type,
  output = c("grouping", "numeric", "percentage")
)

Arguments

age

Input integer value to indicate the age of the patient.

sex

Input characters (female, male) to indicate the sex of the patient.

female

male

have_dyspnoea

Input characters (no, yes) to indicate if the patient only has dyspnoea symptoms.

no stands for not having dyspnoea symptoms.

yes stands for having dyspnoea symptoms.

chest_pain_type

Input characters (no chest pain, typical, atypical, nonanginal) to indicate the chest pain characteristics of the patient.

no chest pain stands for the patient having no chest pain.

typical stands for the patient having typical chest pain.

atypical stands for the patient having atypical chest pain.

nonanginal stands for the patient having nonanginal or non-specific chest pain.

output

Input text to indicate the how pre-test probability results be expressed Default: c("grouping", "numeric", "percentage")

grouping means the PTP will be expressed as Low, Intermediate and High.

low if PTP is less than 5%.

intermediate if PTP is in between 5% to 15%.

high if PTP is more than 15%.

numeric means the PTP will be expressed as an integer probability (0-100).

percentage means the PTP will be expressed as percentage text (0-100%).

Value

An integer, percentage or category representing the patient's PTP for obstructive CAD based on the ESC 2019 guidelines. See parameter option output for more information.

Details

The predictive model used to create the guidelines are based on patients from European countries with low cardiovascular disease (CVD) risk.

If the patient has both dyspnoea and a particular chest pain type (typical, atypical, nonanginal), The chest pain type will take precedence over dyspnoea.

Examples

# 35 year old female with typical chest pain
calculate_esc_2019_ptp(
    age = 35,
    sex = "female",
    have_dyspnoea = "no",
    chest_pain_type = "typical",
    output = "percentage"
)
#> [1] "5%"

# 75 year old male with only dyspnoea
calculate_esc_2019_ptp(
    age = 75,
    sex = "male",
    have_dyspnoea = "yes",
    chest_pain_type = "no chest pain",
    output = "percentage"
)
#> [1] "32%"