Calculate ESC 2024 PTP for obstructive CAD

This function returns a patient's pre-test Probability (PTP) of obstructive coronary artery disease (CAD) based on the European Society of Cardiology (ESC) 2024 guidelines.

Usage

calculate_esc_2024_fig_4_ptp_simplfied(
  age,
  sex,
  symptom_score,
  num_of_rf,
  output = c("grouping", "numeric", "percentage")
)

Arguments

age

Input numeric value to indicate the age of the patient.

sex

Input characters (female, male) to indicate the sex of the patient.

female

male

symptom_score

An integer indicating the symptom score of the patient. This value can be calculated via the calculate_esc_2024_symptom_score

num_of_rf

An integer indicating the number of risk factors the patient has. This value can be calculated via the calculate_esc_2024_num_of_rf Risk factors are:

having a family history of CAD.

having a smoking history (current and past smoker).

having dyslipidemia.

having hypertension.

having diabetes.

output

Input text to indicate the how pre-test probability results be expressed Default: c("grouping", "numeric", "percentage")

grouping means the PTP will be expressed as Low, Intermediate and High.

very low if PTP is less than or equal to 5

low if PTP is in between 6

moderate if PTP is more than 15

numeric means the PTP will be expressed as an integer probability (0-100).

percentage means the PTP will be expressed as percentage text (0-100%).

Value

An integer, percentage or category representing the patient's PTP for obstructive CAD based on the ESC 2024 guidelines. See parameter option `output` for more information.

Examples

# 30 female with symptom score of 0 and 0 risk factors
calculate_esc_2024_fig_4_ptp_simplfied(
  age = 30,
  sex = "female",
  symptom_score = 0,
  num_of_rf = 0,
  output = "percentage"
)
#> [1] "0%"