# https://github.com/r-lib/pkgdown/issues/2704

Calculate 2021 PRECISE Clinical PTP for obstructive CAD

This function returns a patient's pre-test probability (PTP) of obstructive coronary artery disease based on the 2021 Predictive Risk scorE for CAD In Southeast Asians with chEst pain (PRECISE) clinical model.

Usage

calculate_precise_2021_clinical_ptp(
  age,
  sex,
  chest_pain_type,
  have_neck_radiation,
  have_diabetes,
  have_hypertension,
  smoking_history_type,
  have_q_waves,
  have_st_t_changes
)

Arguments

age

Input integer value to indicate the age of the patient.

sex

Input characters (female, male) to indicate the sex of the patient.

female

male

chest_pain_type

Input characters (typical, atypical, nonanginal) to indicate the chest pain characteristics of the patient.

typical stands for the patient having typical chest pain.

atypical stands for the patient having atypical chest pain.

nonanginal stands for the patient having nonanginal or non-specific chest pain.

have_neck_radiation

Input characters (no, yes) to indicate if the patient has chest pain radiating to the neck.

no stands for not chest pain radiating to the neck.

yes stands for having chest pain radiating to the neck.

have_diabetes

Input characters (no, yes) to indicate if the patient has diabetes.

no stands for not having diabetes.

yes stands for having diabetes.

have_hypertension

Input characters (no, yes) to indicate if the patient has hypertension.

no stands for not having hypertension.

yes stands for having hypertension.

smoking_history_type

Input characters (current, past, none) to indicate the smoking history of the patient.

current stands for being a current smoker.

past stands for being a past smoker.

none stands for not having a smoking history (non-smoker).

have_q_waves

Input characters (no, yes) to indicate if the patient has Q waves on electrocardiogram (ECG).

no stands for the patient not having Q waves on ECG.

yes stands for the patient having Q waves on ECG.

have_st_t_changes

Input characters (no, yes) to indicate if the patient has ST-T changes on electrocardiogram (ECG).

no stands for the patient not having ST-T changes on ECG.

yes stands for the patient having ST-T changes on ECG.

Value

A numeric value representing the patient's PTP for obstructive CAD based on the 2021 Predictive Risk scorE for CAD In Southeast Asians with chEst pain (PRECISE) clinical model.

Details

The predictive model is based on patients a mixed Asian cohort within Singapore with stable chest pain.

Examples

# 40 year old female with typical chest pain
# radiating to the neck, has diabetes
# but no hypertension and a non-smoker.
# She has Q waves but no ST-T changes on ECG.

calculate_precise_2021_clinical_ptp(
    age = 40,
    sex = "female",
    chest_pain_type = "typical",
    have_neck_radiation = "yes",
    have_diabetes = "yes",
    have_hypertension = "no",
    smoking_history_type = "none",
    have_q_waves = "yes",
    have_st_t_changes = "no"
)
#> [1] 0.2233933