# https://github.com/r-lib/pkgdown/issues/2704

Check If Positive

Check if the input variable is a positive number

Usage

check_if_positive(
  x,
  allow_na = TRUE,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)

Arguments

x

Input variable to check if it is positive number

allow_na

Input boolean to determine if NA or NaN is allowed. Default: TRUE

arg

An argument name as a string. This argument will be mentioned in error messages as the input that is at the origin of a problem.

call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

Value

The variable itself or an error message if variable is not positive

See also