Calculate ESC 2024 PTP for obstructive CAD

This function returns a patient's pre-test Probability (PTP) of obstructive coronary artery disease (CAD) based on the European Society of Cardiology (ESC) 2024 guidelines.

Usage

calculate_esc_2024_fig_4_ptp(
  age,
  sex,
  chest_pain_type,
  have_dyspnea,
  have_family_history,
  have_smoking_history,
  have_dyslipidemia,
  have_hypertension,
  have_diabetes,
  allow_na_symptom_score = TRUE,
  max_na_num_of_rf = 0,
  output = c("grouping", "numeric", "percentage")
)

Arguments

age

Input numeric value to indicate the age of the patient.

sex

Input characters (female, male) to indicate the sex of the patient.

female

male

chest_pain_type

Input characters (no chest pain, typical, atypical, nonanginal) to indicate the chest pain characteristics of the patient.

no chest pain stands for the patient having no chest pain.

typical stands for the patient having typical chest pain.

atypical stands for the patient having atypical chest pain.

nonanginal stands for the patient having nonanginal or non-specific chest pain.

have_dyspnea

Input characters (no, yes) to indicate if the patient only has dyspnea symptoms.

no stands for not having dyspnea symptoms.

yes stands for having dyspnea symptoms.

have_family_history

Input characters (no, yes) to indicate if the patient only has a family history of CAD.

no stands for not having a family history of CAD.

yes stands for having a family history of CAD.

have_smoking_history

Input characters (no, yes) to indicate if the patient only has a smoking history (current or past smoker).

no stands for not having a smoking history (non-smoker).

yes stands for having a smoking history (current or past smoker).

have_dyslipidemia

Input characters (no, yes) to indicate if the patient only has dyslipidemia.

no stands for not having dyslipidemia.

yes stands for having a dyslipidemia.

have_hypertension

Input characters (no, yes) to indicate if the patient only has hypertension.

no stands for not having hypertension.

yes stands for having a hypertension.

have_diabetes

Input characters (no, yes) to indicate if the patient only has diabetes.

no stands for not having diabetes.

yes stands for having a diabetes.

allow_na_symptom_score

A logical evaluating to TRUE or FALSE indicating whether we can allow `chest_pain_type` or `have_dyspnea` to be NA when calculating the score

max_na_num_of_rf

Input integer 0 to 5 to indicate the maximum number of missing risk factors to tolerate before outputting an NA. Default: 0

output

Input text to indicate the how pre-test probability results be expressed Default: c("grouping", "numeric", "percentage")

grouping means the PTP will be expressed as Low, Intermediate and High.

very low if PTP is less than or equal to 5

low if PTP is in between 6

moderate if PTP is more than 15

numeric means the PTP will be expressed as an integer probability (0-100).

percentage means the PTP will be expressed as percentage text (0-100%).

Value

An integer indicating the symptom score of the patient. Default: TRUE

An integer, percentage or category representing the patient's PTP for obstructive CAD based on the ESC 2024 guidelines. See parameter option `output` for more information.

Examples

# 30 female with symptom score of 0 and 0 risk factors
calculate_esc_2024_fig_4_ptp(
  age = 30,
  sex = "female",
  chest_pain_type = "no chest pain",
  have_dyspnea = "no",
  have_family_history = "no",
  have_smoking_history = "no",
  have_dyslipidemia = "no",
  have_hypertension = "no",
  have_diabetes = "no",
  allow_na_symptom_score = TRUE,
  max_na_num_of_rf = 0,
  output = "percentage"
)
#> [1] "0%"