# https://github.com/r-lib/pkgdown/issues/2704

Calculate Symptom Score (ESC 2024)

A function used to calculate the symptom score of the patient. This is used to calculate the pretest probability of coronary artery disease (CAD) based on the ESC 2024 guidelines.

Usage

calculate_esc_2024_symptom_score(
  chest_pain_type,
  have_dyspnoea,
  allow_na = TRUE
)

Arguments

chest_pain_type

Input characters (no chest pain, typical, atypical, nonanginal) to indicate the chest pain characteristics of the patient.

no chest pain stands for the patient having no chest pain.

typical stands for the patient having typical chest pain.

atypical stands for the patient having atypical chest pain.

nonanginal stands for the patient having nonanginal or non-specific chest pain.

have_dyspnoea

Input characters (no, yes) to indicate if the patient only has dyspnoea symptoms.

no stands for not having dyspnoea symptoms.

yes stands for having dyspnoea symptoms.

allow_na

A logical evaluating to TRUE or FALSE indicating whether we can allow `chest_pain_type` or `have_dyspnoea` to be NA when calculating the score. Default: TRUE

Value

An integer indicating the symptom score of the patient. It can also be NA if both chest_pain_type and have_dyspnoea are NA. Patients with both nonanginal chest pain and dyspnoea will be given a score of 2

Examples

calculate_esc_2024_symptom_score(
  chest_pain_type = "nonanginal",
  have_dyspnoea = "yes",
  allow_na = TRUE
)
#> [1] 2

calculate_esc_2024_symptom_score(
  chest_pain_type = "nonanginal",
  have_dyspnoea = NA,
  allow_na = FALSE
)
#> [1] NA

calculate_esc_2024_symptom_score(
  chest_pain_type = "nonanginal",
  have_dyspnoea = NA,
  allow_na = TRUE
)
#> [1] 1