Calculate 2017 PROMISE Minimal-Risk Score for obstructive CAD

This function returns a symptomatic (have chest pain or dyspnea) patient's minimal risk score for obstructive coronary artery disease based on the 2017 PROMISE Minimal-Risk Score.

Usage

calculate_prms_2017_ptp(
  age,
  sex,
  hdl_mg_dl,
  is_minority_ethnicity,
  has_diabetes,
  has_hypertension,
  has_dyslipidemia,
  has_smoking_history,
  has_family_history_of_cad,
  has_stress_symptoms = NA
)

Arguments

age

Input integer to indicate the age of the patient.

sex

Input integer 0 or 1 to indicate the sex of the patient.

0 stands for Female

1 stands for Male

hdl_mg_dl

Input positive numeric value to indicate the patient's high-density lipoprotein (HDL) in \(mg/dL\).

is_minority_ethnicity

Input integer 0 or 1 to indicate if the patient is from a racial or minority ethnicity (or patient is not a non-Hispanic/Latino White).

0 stands for patient is a non-Hispanic/Latino White.

1 stands for patient is not a non-Hispanic/Latino White. E.g. Blacks, Asians, etc.

has_diabetes

Input integer 0 or 1 to indicate if the patient has diabetes.

0 stands for not having diabetes.

1 stands for having diabetes.

has_hypertension

Input integer 0 or 1 to indicate if the patient has hypertension.

0 stands for not having hypertension.

1 stands for having hypertension.

has_dyslipidemia

Input integer 0 or 1 to indicate if the patient has dyslipidemia.

0 stands for not having dyslipidemia.

1 stands for having dyslipidemia.

has_smoking_history

Input integer 0 or 1 to indicate if the patient has a smoking history (current or past smoker).

0 stands for not having a smoking history (non-smoker).

1 stands for having a smoking history (current or past smoker).

has_family_history_of_cad

Input integer 0 or 1 to indicate if the patient has a family history of CAD.

0 stands for not having a family history of CAD.

1 stands for having a family history of CAD.

has_stress_symptoms

Input integer 0 or 1 to indicate if the patient has symptoms related to physical or mental stress. It can be set to NA if the patient results are inconclusive or have not taken any stress test such as exercise treadmill testing, stress echocardiography, or stress nuclear imaging.

0 stands for no symptoms (negative results) related to physical or mental stress.

1 stands for having symptoms (positive results) related to physical or mental stress.

NA stands for inconclusive results or patient has not taken any stress test

Default: NA

Value

A numeric value representing the patient's minimal risk score for obstructive CAD based on the 2017 PROMISE Minimal-Risk Score.

Details

The predictive model is based on CCTA images from 4632 patients in the Prospective Multicenter imaging Study for Evaluation of Chest Pain (PROMISE) trial.

Examples

# 50 year old white female with chest pain
# a medical history of hypertension, and a
# high-density lipoprotein cholesterol level of 70 mg/dL
calculate_prms_2017_ptp(
    age = 50,
    sex = 0,
    hdl_mg_dl = 70,
    is_minority_ethnicity = 0,
    has_diabetes = 0,
    has_hypertension = 1,
    has_dyslipidemia = 0,
    has_smoking_history = 0,
    has_family_history_of_cad = 0,
    has_stress_symptoms = 0
)
#> [1] 0.710744

# 40 year old non-white male with chest pain
# a medical history of diabetes, unknown stress symptoms and a
# high-density lipoprotein cholesterol level of 65 mg/dL
calculate_prms_2017_ptp(
    age = 40,
    sex = 1,
    hdl_mg_dl = 70,
    is_minority_ethnicity = 1,
    has_diabetes = 1,
    has_hypertension = 0,
    has_dyslipidemia = 0,
    has_smoking_history = 0,
    has_family_history_of_cad = 0,
    has_stress_symptoms = NA
)
#> [1] 0.6974111