# https://github.com/r-lib/pkgdown/issues/2704

Calculate 2017 PROMISE Minimal-Risk Score for obstructive CAD

This function returns a symptomatic (have chest pain or dyspnoea) patient's minimal risk score for obstructive coronary artery disease based on the 2017 PROMISE Minimal-Risk Score.

Usage

calculate_prms_2017_ptp(
  age,
  sex,
  hdl_mg_dl,
  is_minority_ethnicity,
  have_diabetes,
  have_hypertension,
  have_dyslipidemia,
  have_smoking_history,
  have_family_history,
  have_stress_symptoms = NA
)

Arguments

age

Input numeric value to indicate the age of the patient.

sex

Input characters (female, male) to indicate the sex of the patient.

female

male

hdl_mg_dl

Input positive numeric value to indicate the patient's high-density lipoprotein (HDL) in \(mg/dL\).

is_minority_ethnicity

Input characters (no, yes) to indicate if the patient is from a racial or minority ethnicity (or patient is not a non-Hispanic/Latino White).

no stands for patient is a non-Hispanic/Latino White.

yes stands for patient is not a non-Hispanic/Latino White. E.g. Blacks, Asians, etc.

have_diabetes

Input characters (no, yes) to indicate if the patient has diabetes.

no stands for not having diabetes.

yes stands for having diabetes.

have_hypertension

Input characters (no, yes) to indicate if the patient has hypertension.

no stands for not having hypertension.

yes stands for having hypertension.

have_dyslipidemia

Input characters (no, yes) to indicate if the patient has dyslipidemia.

no stands for not having dyslipidemia.

yes stands for having dyslipidemia.

have_smoking_history

Input characters (no, yes) to indicate if the patient has a smoking history (current or past smoker).

no stands for not having a smoking history (non-smoker).

yes stands for having a smoking history (current or past smoker).

have_family_history

Input characters (no, yes) to indicate if the patient has a family history of CAD.

no stands for not having a family history of CAD.

yes stands for having a family history of CAD.

have_stress_symptoms

Input characters (no, yes) to indicate if the patient has symptoms related to physical or mental stress. It can be set to NA if the patient results are inconclusive or have not taken any stress test such as exercise treadmill testing, stress echocardiography, or stress nuclear imaging.

no stands for no symptoms (negative results) related to physical or mental stress.

yes stands for having symptoms (positive results) related to physical or mental stress.

NA stands for inconclusive results or patient has not taken any stress test

Default: NA

Value

A numeric value representing the patient's minimal risk score for obstructive CAD based on the 2017 PROMISE Minimal-Risk Score.

Details

The predictive model is based on CCTA images from 4632 patients in the Prospective Multicenter imaging Study for Evaluation of Chest Pain (PROMISE) trial.

Examples

# 50 year old white female with chest pain
# a medical history of hypertension, and a
# high-density lipoprotein cholesterol level of 70 mg/dL
calculate_prms_2017_ptp(
    age = 50,
    sex = "female",
    hdl_mg_dl = 70,
    is_minority_ethnicity = "no",
    have_diabetes = "no",
    have_hypertension = "yes",
    have_dyslipidemia = "no",
    have_smoking_history = "no",
    have_family_history = "no",
    have_stress_symptoms = "no"
)
#> [1] 0.710744

# 40 year old non-white male with chest pain
# a medical history of diabetes, unknown stress symptoms and a
# high-density lipoprotein cholesterol level of 70 mg/dL
calculate_prms_2017_ptp(
    age = 40,
    sex = "male",
    hdl_mg_dl = 70,
    is_minority_ethnicity = "yes",
    have_diabetes = "yes",
    have_hypertension = "no",
    have_dyslipidemia = "no",
    have_smoking_history = "no",
    have_family_history = "no",
    have_stress_symptoms = NA
)
#> [1] 0.6974111