# https://github.com/r-lib/pkgdown/issues/2704

Calculate 2019 Reeh Basic PTP for obstructive CAD

This function returns a patient's pre-test probability (PTP) of obstructive coronary artery disease based on the 2019 Reeh et. al. basic model.

Usage

calculate_reeh_2019_basic_ptp(age, sex, symptom_type)

Arguments

age

Input integer value to indicate the age of the patient.

sex

Input characters (female, male) to indicate the sex of the patient.

female

male

symptom_type

Input characters (typical, atypical, nonanginal, dyspnoea) to indicate the symptom characteristics of the patient.

typical stands for the patient having typical chest pain.

atypical stands for the patient having atypical chest pain.

nonanginal stands for the patient having nonanginal or non-specific chest pain.

dyspnoea stands for the patient having dyspnoea.

Value

A numeric value representing the patient's PTP for obstructive CAD based on the 2019 Reeh et. al. basic model.

Details

The predictive model is based on 3903 patients free of CAD and heart failure and suspected of angina, who were referred to a single, large, urban university hospital for assessment in 2012–15.

Examples

# 40 year old female with typical chest pain
calculate_reeh_2019_basic_ptp(
    age = 40,
    sex = "female",
    symptom_type = "typical"
)
#> [1] 0.05578231