Calculate 2020 Winther Basic PTP for obstructive CAD

This function returns a patient's pre-test probability (PTP) of obstructive coronary artery disease based on the 2020 Winther et. al. basic model (Basic_PTP).

Usage

calculate_winther_2020_basic_ptp(age, sex, chest_pain)

Arguments

age

Input integer to indicate the age of the patient.

sex

Input integer 0 or 1 to indicate the sex of the patient.

0 stands for Female

1 stands for Male

chest_pain

Input integer 1 to 3 to indicate the chest pain characteristics of the patient.

1 stands for the patient having typical chest pain.

2 stands for the patient having atypical chest pain or dyspnea.

3 stands for the patient having non-anginal or non-specific chest pain.

Value

A numeric value representing the patient's PTP for obstructive CAD based on the 2020 Winther et. al. basic model (Basic_PTP).

Details

The predictive model is based on > 40000 symptomatic patients from 2008 to 2017 from 13 hospitals in Western Denmark. These patients are registered under the Western Denmark Heart Registry.

Examples

# 40 year old Male with typical chest pain
calculate_winther_2020_basic_ptp(
    age = 40,
    sex = 1,
    chest_pain = 1
)
#> [1] 0.1430237

# 40 year old Male with non-anginal chest pain
calculate_winther_2020_basic_ptp(
    age = 40,
    sex = 1,
    chest_pain = 3
)
#> [1] 0.02218292